requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blosc2==2.4.0
bottle==0.12.25
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
ezodf==0.3.2
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
h5py==3.11.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lightning==2.1.1
lightning-utilities==0.10.0
lit==17.0.6
lxml==4.9.4
Mako==1.3.3
Markdown==3.5.1
markdown-it-py==3.0.0
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
namex==0.0.8
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
ndindex==1.7
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.8
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
optuna==3.6.1
optuna-dashboard==0.15.1
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandas-ods-reader==0.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
Pillow==10.0.1
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==0.8.5
pytorchtools==0.0.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.16.2
rsa==4.9
schedulefree==1.2.7
scikit-image==0.24.0
scikit-learn==1.5.1
scikit-plot==0.3.7
scipy==1.11.4
Send2Trash==1.8.2
sentry-sdk==1.39.1
setproctitle==1.3.3
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
sympy==1.12
tables==3.9.2
tensorboard==2.15.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2024.6.18
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torcheval==0.0.7
torchmetrics==1.2.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.0.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==2.1.0
wandb==0.16.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.4